+++ libs/langchain-google-common/src/utils/safety.ts
@@ -1,15 +1,5 @@
 import { GoogleLLMResponse } from "../types.js";
 
-export interface GoogleAISafetyHandler {
-  /**
-   * A function that will take a response and return the, possibly modified,
-   * response or throw an exception if there are safety issues.
-   *
-   * @throws GoogleAISafetyError
-   */
-  handle(response: GoogleLLMResponse): GoogleLLMResponse;
-}
-
 export class GoogleAISafetyError extends Error {
   response: GoogleLLMResponse;
 
@@ -21,8 +11,4 @@ export class GoogleAISafetyError extends Error {
 
     this.response = response;
   }
-}
-
-export interface GoogleAISafetyParams {
-  safetyHandler?: GoogleAISafetyHandler;
 }
